Update README.md
Text :"* *NOTE* * This is an experimental Repo which is done for understanding GitHub theme pages and is not in its final stages. The information available in the index.md file is no where near its actual situation." 

the above text has been added, since this repo is not final and I'll make my own portfolio using react.js from scratch.